[CLI-1143] Make inquirer return default answers again
We often add default values for answers inside the "when" function of a
question and then return false. Inquirer will not return those answers
since 1.0.0 as it means they were skipped entirely. To workaround this
we can get our injected default values from the prompt UI reference
that's attached to the promise and holds the answers before they where
filtered by inquirer.